+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `sdwan_system_snmp_profile_parcel` resource and data source
 - Add `sdwan_system_basic_profile_parcel` resource and data source
 - Fix issue with deleted feature profile resource, [link](https://github.com/CiscoDevNet/terraform-provider-sdwan/issues/204)
+- Add `sdwan_transport_wan_vpn_profile_parcel` resource and data source
 
 ## 0.3.7